Many upgrades

+++ source/montecarlo.h
@@ -1,15 +1,9 @@
 #pragma once
 
-#include "constants.h"
-#include "main.h"
 #include "initialization.h"
-#include "rng.h"
-#include "class_tic_toc.h"
-#include <iostream>
 #include <cstring>
-#include "o2.h"
-#include "measures.h"
+#include <iostream>
 
-void metropolis(const std::vector<Node> &Site, struct MC_parameters &MCp, struct H_parameters &Hp, double my_beta);
-double local_HPsi(std::array<O2, NC> &Psi, int ix, int iy, struct H_parameters &Hp, const std::vector<Node> &Site);
-double local_HA(double A, int ix, int iy, int alpha,  struct H_parameters &Hp, const std::vector<Node> &Site);
+void   metropolis(const std::vector<Node> &Site, struct MC_parameters &MCp, struct H_parameters &Hp, double my_beta);
+double local_HPsi(std::array<O2, NC> &Psi, size_t ix, size_t iy, struct H_parameters &Hp, const std::vector<Node> &Site);
+double local_HA(double A, size_t ix, size_t iy, size_t alpha, struct H_parameters &Hp, const std::vector<Node> &Site);
